backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pymysql
openpyxl
python-dateutil
